pyproject.toml
[project]
name = "quant-solver"
version = "0.1.0"
description = "Multi-agent verification pipeline for quant interview math problems from screenshots in any language."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "typer>=0.12",
    "structlog>=24.1",
    "sympy>=1.12",
    "numpy>=1.26",
    "scipy>=1.12",
    "python-dotenv>=1.0",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pillow>=10.0",
]

[project.scripts]
quant-solver = "quant_solver.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quant_solver"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: end-to-end tests that hit the live Anthropic API",
]

[tool.ruff]
line-length = 100
target-version = "py311"
